==> 01_Web-Scraping_Datenexport_Ladekurven.ipynb <==
# Web-Scraping für Datenexport der Ladekurven 

import os
import time
from selenium import webdriver
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.common.by import By
from selenium.webdriver.support import expected_conditions as EC

## Aufruf des Dashboard von Chargecloud über Chrome-Webdriver 

driver = webdriver.Chrome(r'Dateipfad1\chromedriver.exe')
driver.get("https://app.chargecloud.de/login")


## Automatische Anmeldung bei Chargecloud

mandant = WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="login_client"]')))
mandant.send_keys("Mandant")
email = WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="login_email"]')))
email.send_keys("E-Mail")
passwort = WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="login_password"]')))
passwort.send_keys("Passwort")
WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="login_submit"]'))).click()

## Danach manuelles Vorgehen:  Standort --> Details(SZ,WF_H oder WF_R) --> Ladevorgänge --> nächsten Codeabschnitt ausführen 

## Automatisierter Datenexport der Messwerte
 
## Anzahl der Seiten die durchlaufen werden
Seiten = 50

for n in range(Seiten):
## Jeder Codeblöck steht für einen Ladevorgang je Seite   
##    
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="1"]/td[15]/div/a[1]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//td[contains(@id, "emobility_chargepoint_transaction_metervalue__")]/table/tbody/tr/td[4]/div/span'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="export_exportCsv"]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//div[contains(@class, "ui-dialog-titlebar")]/button/span[2]'))).click()
    time.sleep(1)
    id_ = WebDriverWait(driver, 10).until(EC.visibility_of_element_located((By.XPATH, '//*[@id="1"]/td[1]'))).text
    ## Messwerte werden unter Downloads abgespeichert.
    name = r"C:\Users\Benutzer\Downloads\export_messwerte.csv"
    ## Verschiebung der Messwerte und unbenennung in "ID".csv
    os.rename(name, r"Dateipfad\\" + id_ + ".csv")
##    
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="2"]/td[15]/div/a[1]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//td[contains(@id, "emobility_chargepoint_transaction_metervalue__")]/table/tbody/tr/td[4]/div/span'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="export_exportCsv"]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//div[contains(@class, "ui-dialog-titlebar")]/button/span[2]'))).click()
    time.sleep(1)
    id_ = WebDriverWait(driver, 10).until(EC.visibility_of_element_located((By.XPATH, '//*[@id="2"]/td[1]'))).text
    name = r"C:\Users\Benutzer\Downloads\export_messwerte.csv"
    os.rename(name, r"Dateipfad\\" + id_ + ".csv")
    
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="3"]/td[15]/div/a[1]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//td[contains(@id, "emobility_chargepoint_transaction_metervalue__")]/table/tbody/tr/td[4]/div/span'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="export_exportCsv"]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//div[contains(@class, "ui-dialog-titlebar")]/button/span[2]'))).click()
    time.sleep(1)
    id_ = WebDriverWait(driver, 10).until(EC.visibility_of_element_located((By.XPATH, '//*[@id="3"]/td[1]'))).text
    name = r"C:\Users\Benutzer\Downloads\export_messwerte.csv"
    os.rename(name, r"Dateipfad\\" + id_ + ".csv")
    
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="4"]/td[15]/div/a[1]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//td[contains(@id, "emobility_chargepoint_transaction_metervalue__")]/table/tbody/tr/td[4]/div/span'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="export_exportCsv"]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//div[contains(@class, "ui-dialog-titlebar")]/button/span[2]'))).click()
    time.sleep(1)
    id_ = WebDriverWait(driver, 10).until(EC.visibility_of_element_located((By.XPATH, '//*[@id="4"]/td[1]'))).text
    name = r"C:\Users\Benutzer\Downloads\export_messwerte.csv"
    os.rename(name, r"Dateipfad\\" + id_ + ".csv")
    
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="5"]/td[15]/div/a[1]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//td[contains(@id, "emobility_chargepoint_transaction_metervalue__")]/table/tbody/tr/td[4]/div/span'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="export_exportCsv"]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//div[contains(@class, "ui-dialog-titlebar")]/button/span[2]'))).click()
    time.sleep(1)
    id_ = WebDriverWait(driver, 10).until(EC.visibility_of_element_located((By.XPATH, '//*[@id="5"]/td[1]'))).text
    name = r"C:\Users\Benutzer\Downloads\export_messwerte.csv"
    os.rename(name, r"Dateipfad\\" + o + ".csv")
    
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="6"]/td[15]/div/a[1]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//td[contains(@id, "emobility_chargepoint_transaction_metervalue__")]/table/tbody/tr/td[4]/div/span'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="export_exportCsv"]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//div[contains(@class, "ui-dialog-titlebar")]/button/span[2]'))).click()
    time.sleep(1)
    id_ = WebDriverWait(driver, 10).until(EC.visibility_of_element_located((By.XPATH, '//*[@id="6"]/td[1]'))).text
    name = r"C:\Users\Benutzer\Downloads\export_messwerte.csv"
    os.rename(name, r"Dateipfad\\" + id_ + ".csv")
    
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="7"]/td[15]/div/a[1]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//td[contains(@id, "emobility_chargepoint_transaction_metervalue__")]/table/tbody/tr/td[4]/div/span'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="export_exportCsv"]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//div[contains(@class, "ui-dialog-titlebar")]/button/span[2]'))).click()
    time.sleep(1)
    id_ = WebDriverWait(driver, 10).until(EC.visibility_of_element_located((By.XPATH, '//*[@id="7"]/td[1]'))).text
    name = r"C:\Users\Benutzer\Downloads\export_messwerte.csv"
    os.rename(name, r"Dateipfad\\" + id_ + ".csv")
    
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="8"]/td[15]/div/a[1]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//td[contains(@id, "emobility_chargepoint_transaction_metervalue__")]/table/tbody/tr/td[4]/div/span'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="export_exportCsv"]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//div[contains(@class, "ui-dialog-titlebar")]/button/span[2]'))).click()
    time.sleep(1)
    id_ = WebDriverWait(driver, 10).until(EC.visibility_of_element_located((By.XPATH, '//*[@id="8"]/td[1]'))).text
    name = r"C:\Users\Benutzer\Downloads\export_messwerte.csv"
    os.rename(name, r"Dateipfad\\" + id_ + ".csv")
    
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="9"]/td[15]/div/a[1]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//td[contains(@id, "emobility_chargepoint_transaction_metervalue__")]/table/tbody/tr/td[4]/div/span'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="export_exportCsv"]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//div[contains(@class, "ui-dialog-titlebar")]/button/span[2]'))).click()
    time.sleep(1)
    id_ = WebDriverWait(driver, 10).until(EC.visibility_of_element_located((By.XPATH, '//*[@id="9"]/td[1]'))).text
    name = r"C:\Users\Benutzer\Downloads\export_messwerte.csv"
    os.rename(name, r"Dateipfad\\" + id_ + ".csv")
    
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="10"]/td[15]/div/a[1]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//td[contains(@id, "emobility_chargepoint_transaction_metervalue__")]/table/tbody/tr/td[4]/div/span'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="export_exportCsv"]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//div[contains(@class, "ui-dialog-titlebar")]/button/span[2]'))).click()
    time.sleep(1)
    id_ = WebDriverWait(driver, 10).until(EC.visibility_of_element_located((By.XPATH, '//*[@id="10"]/td[1]'))).text
    name = r"C:\Users\Benutzer\Downloads\export_messwerte.csv"
    os.rename(name, r"Dateipfad\\" + id_ + ".csv")
    
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="11"]/td[15]/div/a[1]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//td[contains(@id, "emobility_chargepoint_transaction_metervalue__")]/table/tbody/tr/td[4]/div/span'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="export_exportCsv"]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//div[contains(@class, "ui-dialog-titlebar")]/button/span[2]'))).click()
    time.sleep(1)
    id_ = WebDriverWait(driver, 10).until(EC.visibility_of_element_located((By.XPATH, '//*[@id="11"]/td[1]'))).text
    name = r"C:\Users\Benutzer\Downloads\export_messwerte.csv"
    os.rename(name, r"Dateipfad\\" + id_ + ".csv")
    
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="12"]/td[15]/div/a[1]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//td[contains(@id, "emobility_chargepoint_transaction_metervalue__")]/table/tbody/tr/td[4]/div/span'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="export_exportCsv"]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//div[contains(@class, "ui-dialog-titlebar")]/button/span[2]'))).click()
    time.sleep(1)
    id_ = WebDriverWait(driver, 10).until(EC.visibility_of_element_located((By.XPATH, '//*[@id="12"]/td[1]'))).text
    name = r"C:\Users\Benutzer\Downloads\export_messwerte.csv"
    os.rename(name, r"Dateipfad\\" + id_ + ".csv")
    
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="13"]/td[15]/div/a[1]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//td[contains(@id, "emobility_chargepoint_transaction_metervalue__")]/table/tbody/tr/td[4]/div/span'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="export_exportCsv"]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//div[contains(@class, "ui-dialog-titlebar")]/button/span[2]'))).click()
    time.sleep(1)
    id_ = WebDriverWait(driver, 10).until(EC.visibility_of_element_located((By.XPATH, '//*[@id="13"]/td[1]'))).text
    name = r"C:\Users\Benutzer\Downloads\export_messwerte.csv"
    os.rename(name, r"Dateipfad\\" + id_ + ".csv")
    
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="14"]/td[15]/div/a[1]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//td[contains(@id, "emobility_chargepoint_transaction_metervalue__")]/table/tbody/tr/td[4]/div/span'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//*[@id="export_exportCsv"]'))).click()
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//div[contains(@class, "ui-dialog-titlebar")]/button/span[2]'))).click()
    time.sleep(1)
    id_ = WebDriverWait(driver, 10).until(EC.visibility_of_element_located((By.XPATH, '//*[@id="14"]/td[1]'))).text
    name = r"C:\Users\Benutzer\Downloads\export_messwerte.csv"
    os.rename(name, r"Dateipfad\\" + id_ + ".csv")
    
    ##Wechsel auf nächste Seite
    WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.XPATH, '//td[contains(@id, "next_emobility_chargepoint_transaction_location__")]/span'))).click()
